refactor(rfc-9110-rules): rename utils files
