+++ packages/rfc-9110-rules/src/rules/range-requests/content-range/utils.ts
