+++ packages/rfc-9110-rules/src/rules/range-requests/range/utils.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { parseRangeHeader } from './parse-range-header';
+import { parseRangeHeader } from './utils';
 
 describe('parseRangeHeader', () => {
   it('parses a valid single range', () => {
