+++ packages/rfc-9110-rules/src/rules/range-requests/range/utils.ts
